requirements.txt
selenium
Pillow
torch
numpy
opencv-python
tetris-gymnasium